Move delay logic to Step.Loading
