+++ client/landing/stepper/declarative-flow/internals/steps-repository/processing-step/index.tsx
@@ -7,13 +7,13 @@ import {
 	HUNDRED_YEAR_DOMAIN_TRANSFER,
 	isAnyHostingFlow,
 	isNewsletterFlow,
+	Step,
 } from '@automattic/onboarding';
 import { useSelect, useDispatch } from '@wordpress/data';
 import { useI18n } from '@wordpress/react-i18n';
 import { useEffect, useState, useRef } from 'react';
 import DocumentHead from 'calypso/components/data/document-head';
 import Loading from 'calypso/components/loading';
-import { StepContainerV2Loading } from 'calypso/components/loading/StepContainerV2Loading';
 import availableFlows from 'calypso/landing/stepper/declarative-flow/registered-flows';
 import { useRecordSignupComplete } from 'calypso/landing/stepper/hooks/use-record-signup-complete';
 import { ONBOARD_STORE, SITE_STORE } from 'calypso/landing/stepper/stores';
@@ -225,11 +225,7 @@ const ProcessingStep: React.FC< ProcessingStepProps > = function ( props ) {
 		return (
 			<>
 				<DocumentHead title={ __( 'Processing' ) } />
-				<StepContainerV2Loading
-					title={ getCurrentMessage() }
-					progress={ progress }
-					delay={ 1000 }
-				/>
+				<Step.Loading title={ getCurrentMessage() } progress={ progress } delay={ 1000 } />
 			</>
 		);
 	}
