+++ client/landing/stepper/declarative-flow/internals/steps-repository/unified-plans/index.tsx
@@ -13,12 +13,12 @@ import {
 	NEWSLETTER_FLOW,
 	START_WRITING_FLOW,
 	useStepPersistedState,
+	Step,
 } from '@automattic/onboarding';
 import { useDispatch, useSelect, useDispatch as useWPDispatch } from '@wordpress/data';
 import { useState } from 'react';
 import { useQueryTheme } from 'calypso/components/data/query-theme';
 import Loading from 'calypso/components/loading';
-import { StepContainerV2Loading } from 'calypso/components/loading/StepContainerV2Loading';
 import { useQuery } from 'calypso/landing/stepper/hooks/use-query';
 import { useSite } from 'calypso/landing/stepper/hooks/use-site';
 import { useSiteSlug } from 'calypso/landing/stepper/hooks/use-site-slug';
@@ -196,7 +196,7 @@ const PlansStepAdaptor: StepType< {
 	const isUsingStepContainerV2 = shouldUseStepContainerV2( props.flow );
 
 	if ( isLoadingSelectedTheme ) {
-		return isUsingStepContainerV2 ? <StepContainerV2Loading /> : <Loading />;
+		return isUsingStepContainerV2 ? <Step.Loading /> : <Loading />;
 	}
 
 	return (
