+++ client/my-sites/checkout/checkout-thank-you/pending/index.tsx
@@ -2,12 +2,12 @@ import page from '@automattic/calypso-router';
 import { getUrlParts } from '@automattic/calypso-url';
 import { CheckoutErrorBoundary } from '@automattic/composite-checkout';
 import { localizeUrl } from '@automattic/i18n-utils';
+import { Step } from '@automattic/onboarding';
 import { useShoppingCart } from '@automattic/shopping-cart';
 import { AUTO_RENEWAL } from '@automattic/urls';
 import { useTranslate } from 'i18n-calypso';
 import React, { useState, useEffect, useRef } from 'react';
 import Loading from 'calypso/components/loading';
-import { StepContainerV2Loading } from 'calypso/components/loading/StepContainerV2Loading';
 import Main from 'calypso/components/main';
 import { shouldUseStepContainerV2 } from 'calypso/landing/stepper/declarative-flow/helpers/should-use-step-container-v2';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
@@ -31,6 +31,7 @@ import type {
 	OrderTransactionSuccess,
 } from 'calypso/state/selectors/get-order-transaction';
 import type { CalypsoDispatch } from 'calypso/state/types';
+
 import './style.scss';
 
 interface CheckoutPendingProps {
@@ -94,7 +95,7 @@ function CheckoutPending( {
 	} );
 
 	const content = shouldUseStepContainerV2( getSignupCompleteFlowName() ) ? (
-		<StepContainerV2Loading title={ headingText } delay={ 2000 } />
+		<Step.Loading title={ headingText } delay={ 2000 } />
 	) : (
 		<Main className="checkout-thank-you__pending">
 			<Loading className="checkout__pending-content" title={ headingText } />
