+++ packages/onboarding/src/step-container-v2/wireframes/Loading/Loading.tsx
@@ -1,17 +1,33 @@
 import { ProgressBar } from '@wordpress/components';
-import { ReactNode } from 'react';
+import { ReactNode, useEffect, useState } from 'react';
 import { Heading } from '../../components/Heading/Heading';
 import { StepContainerV2 } from '../../components/StepContainerV2/StepContainerV2';
 import { TopBar } from '../../components/TopBar/TopBar';
 
 import './style.scss';
 
-export const Loading = ( { title, progress }: { title?: ReactNode; progress?: number } ) => {
+interface LoadingProps {
+	title?: ReactNode;
+	progress?: number;
+	delay?: number;
+}
+
+export const Loading = ( { title, progress, delay = 0 }: LoadingProps ) => {
+	const [ shouldDisplayTitle, setShouldDisplayTitle ] = useState( delay === 0 );
+
+	useEffect( () => {
+		const timeout = setTimeout( () => {
+			setShouldDisplayTitle( true );
+		}, delay );
+
+		return () => clearTimeout( timeout );
+	}, [ delay ] );
+
 	return (
 		<StepContainerV2>
 			<TopBar />
 			<div className="step-container-v2--loading">
-				{ title && (
+				{ title && shouldDisplayTitle && (
 					<div className="step-container-v2--loading__heading-wrapper">
 						<div className="step-container-v2--loading__heading">
 							<Heading text={ title } size="small" align="center" />
